requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
Flask==0.12.3
gunicorn==19.7.1
html5lib==1.0.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==4.3.0
python-magic==0.4.15
requests==2.18.4
six==1.11.0
Werkzeug==0.13
sentry-sdk[flask]==0.7.2
blinker==1.4
python-datauri==0.2.5
CairoSVG==2.1.2
python-dotenv
mypy